Add support for mouse
A _lot_ had to be rebuilt to arrange for this; several of the includes
were poorly configured, and build issues arose. Now, the window has
a reference to the camera.

The Application and Window classes are somewhat improved, while the Camera
class has deteriorated; the camera will have be wrapped in a controller later.

Some changes here should admittedly have been separated, such as merging the
CMakeLists files and creating a source file for the window.